Event-Mate-Track-main/EventMate/requirements.txt
# Core dependencies
uagents>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# AI/ML dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Optional but recommended for production
# uvicorn[standard]>=0.24.0  # Uncomment for production server
# python-multipart>=0.0.6    # Uncomment if handling file uploads 